requirements.txt
requests>=2.31.0
pandas>=2.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
gspread>=6.0.0
google-auth>=2.20.0
